+++ submission/submission-security/src/main/kotlin/ebi/ac/uk/security/service/SecurityService.kt
@@ -14,6 +14,9 @@ import ebi.ac.uk.extended.events.SecurityNotification
 import ebi.ac.uk.extended.events.SecurityNotificationType.ACTIVATION
 import ebi.ac.uk.extended.events.SecurityNotificationType.ACTIVATION_BY_EMAIL
 import ebi.ac.uk.extended.events.SecurityNotificationType.PASSWORD_RESET
+import ebi.ac.uk.io.FileUtils
+import ebi.ac.uk.io.RWXRWX___
+import ebi.ac.uk.io.RWX__X___
 import ebi.ac.uk.model.User
 import ebi.ac.uk.security.integration.components.ISecurityService
 import ebi.ac.uk.security.integration.exception.ActKeyNotFoundException
@@ -21,6 +24,8 @@ import ebi.ac.uk.security.integration.exception.LoginException
 import ebi.ac.uk.security.integration.exception.UserAlreadyRegister
 import ebi.ac.uk.security.integration.exception.UserNotFoundByEmailException
 import ebi.ac.uk.security.integration.exception.UserPendingRegistrationException
+import ebi.ac.uk.security.integration.model.api.FtpMagicFolder
+import ebi.ac.uk.security.integration.model.api.NfsMagicFolder
 import ebi.ac.uk.security.integration.model.api.SecurityUser
 import ebi.ac.uk.security.integration.model.api.UserInfo
 import ebi.ac.uk.security.persistence.getActiveByEmail
@@ -31,13 +36,14 @@ import ebi.ac.uk.security.persistence.getInactiveByEmail
 import ebi.ac.uk.security.util.SecurityUtil
 import org.springframework.transaction.annotation.Transactional
 import uk.ac.ebi.events.service.EventsPublisherService
+import java.nio.file.Path
+import java.nio.file.Paths
 
 @Suppress("TooManyFunctions")
 @Transactional
 open class SecurityService(
     private val userRepository: UserDataRepository,
     private val securityUtil: SecurityUtil,
-    private val magicFolderUtil: MagicFolderUtil,
     private val securityProps: SecurityProperties,
     private val profileService: ProfileService,
     private val captchaVerifier: CaptchaVerifier,
@@ -149,8 +155,26 @@ open class SecurityService(
         val dbUser = userRepository.save(toActivate.apply { activationKey = null; active = true })
         val securityUser = profileService.asSecurityUser(dbUser)
 
-        magicFolderUtil.createMagicFolder(securityUser)
+        createMagicFolder(securityUser)
         return securityUser
+    }
+
+    private fun createMagicFolder(user: SecurityUser) {
+        when (user.magicFolder) {
+            is FtpMagicFolder -> TODO()
+            is NfsMagicFolder -> createNfsMagicFolder(user.email, user.magicFolder)
+        }
+    }
+
+    private fun createNfsMagicFolder(email: String, magicFolder: NfsMagicFolder) {
+        FileUtils.getOrCreateFolder(magicFolder.path.parent, RWX__X___)
+        FileUtils.getOrCreateFolder(magicFolder.path, RWXRWX___)
+        FileUtils.createSymbolicLink(symLinkPath(email), magicFolder.path, RWXRWX___)
+    }
+
+    private fun symLinkPath(userEmail: String): Path {
+        val prefixFolder = userEmail.substring(0, 1).lowercase()
+        return Paths.get("${securityProps.magicDirPath}/$prefixFolder/$userEmail")
     }
 
     private fun asUser(registerRequest: RegisterRequest) = DbUser(
